Parameter to automatically delete leftover or not
- Nextflow parameter added to the main.nf pipeline
  that switches on/off the automatic deletion of
  wgd and i-ADHoRe leftover folders such as
  ks_tmp.
- Default true: it is automatically deleted when
  the pipeline crashes.
- Can be turned off either from the NF
  configuration file or from the command line.
- Note: configuration files must still be updated

+++ main.nf
@@ -17,6 +17,9 @@ params.nThreadsParalogs = 1
 // (will be overriden by job configuration if run on a cluster)
 params.nThreadsOrthologs = 1
 
+// Parameter to automatically delete or not leftover folders at the end of the pipeline
+params.delete_leftover_folders = true
+
 // giving the configuration file through the "input" process section
 configfile = file(params.config)
 
@@ -908,53 +911,57 @@ workflow.onComplete {
     if (LOG) {
         log.info ""
         log.info("Logs folder: logs_${workflow.sessionId.toString().substring(0,8)}")
-        log.info "Cleaning up any temporary files left behind..."
 
-        species_name = file("${configfile}").readLines()[1].split()[2]
-        paralog_dir_path = "${workflow.launchDir}/paralog_distributions/wgd_${species_name}*"
-        ortholog_dir_path = "${workflow.launchDir}/ortholog_distributions/wgd_*"
+        if ( params.delete_leftover_folders == true ) {
+            log.info "Cleaning up any temporary files left behind..."
 
-        // Cleaning both paralog and ortholog directories
-        for ( dir_path : [ paralog_dir_path, ortholog_dir_path ] ) {
-            file(dir_path, type: "dir")
-                .each { wgd_dir ->
-                    // Remove BLAST temporary folder, if any
-                    file("${wgd_dir}/*.blast_tmp", type: "dir")
-                        .each { tmp ->
-                            if ( tmp.exists() == true ) {
-                                result = tmp.deleteDir();
-                                log.info result ? "Deleted: $tmp" : "Can't delete: $tmp"
-                                // Remove associated incomplete BLAST TSV file
-                                file("${wgd_dir}/*.blast.tsv", type: "file")
-                                    .each { tsv ->
-                                        if ( tsv.exists() == true ) {
-                                            result = tsv.delete()
-                                            log.info result ? "Deleted: $tsv" : "Can't delete: $tsv"
+            species_name = file("${configfile}").readLines()[1].split()[2]
+            paralog_dir_path = "${workflow.launchDir}/paralog_distributions/wgd_${species_name}*"
+            ortholog_dir_path = "${workflow.launchDir}/ortholog_distributions/wgd_*"
+
+            // Cleaning both paralog and ortholog directories
+            for ( dir_path : [ paralog_dir_path, ortholog_dir_path ] ) {
+                file(dir_path, type: "dir")
+                    .each { wgd_dir ->
+                        // Remove BLAST temporary folder, if any
+                        file("${wgd_dir}/*.blast_tmp", type: "dir")
+                            .each { tmp ->
+                                if ( tmp.exists() == true ) {
+                                    result = tmp.deleteDir();
+                                    log.info result ? "Deleted: $tmp" : "Can't delete: $tmp"
+                                    // Remove associated incomplete BLAST TSV file
+                                    file("${wgd_dir}/*.blast.tsv", type: "file")
+                                        .each { tsv ->
+                                            if ( tsv.exists() == true ) {
+                                                result = tsv.delete()
+                                                log.info result ? "Deleted: $tsv" : "Can't delete: $tsv"
+                                            }
                                         }
-                                    }
+                                }
                             }
-                        }
-                    // Remove Ks temporary folder, if any
-                    file("${wgd_dir}/*.ks_tmp", type: "dir")
-                        .each { tmp ->
-                            if ( tmp.exists() == true ) {
-                                result = tmp.deleteDir();
-                                log.info result ? "Deleted: $tmp" : "Can't delete: $tmp"
+                        // Remove Ks temporary folder, if any
+                        file("${wgd_dir}/*.ks_tmp", type: "dir")
+                            .each { tmp ->
+                                if ( tmp.exists() == true ) {
+                                    result = tmp.deleteDir();
+                                    log.info result ? "Deleted: $tmp" : "Can't delete: $tmp"
+                                }
                             }
-                        }
-                    // Remove i-ADHoRe temporary folder, if any (applicable only to paralog_distributions)
-                     file("${wgd_dir}/*.ks_anchors_tmp", type: "dir")
-                        .each { tmp ->
-                            if ( tmp.exists() == true ) {
-                                result = tmp.deleteDir();
-                                log.info result ? "Deleted: $tmp" : "Can't delete: $tmp"
+                        // Remove i-ADHoRe temporary folder, if any (applicable only to paralog_distributions)
+                        file("${wgd_dir}/*.ks_anchors_tmp", type: "dir")
+                            .each { tmp ->
+                                if ( tmp.exists() == true ) {
+                                    result = tmp.deleteDir();
+                                    log.info result ? "Deleted: $tmp" : "Can't delete: $tmp"
+                                }
                             }
-                        }
-                }
+                    }
+            }
+
+            log.info "Done."
+            log.info ""
         }
 
-        log.info "Done."
-        log.info ""
         log.info "Pipeline completed at: $workflow.complete taking $workflow.duration"
         log.info "Execution status: ${ workflow.success ? 'OK' : 'failed' }"
         log.info ""
